Split the oversized booking-model test file
AGENTS.md's ~400-line guideline applies to test files too, specifically so
mutation runs stay narrow and fast — the single booking-model.test.ts had
grown to 882 lines covering all of model.ts's exports at once.

Split into three focused suites by concern (child eligibility, capacity,
day-count support), each comfortably under 400 lines, plus a shared
non-test fixtures module (booking-model-fixtures.ts) so the handful of
fixture helpers used across all three aren't duplicated.

No coverage change: mutation testing against the three split files
together still detects 155/165 (100%, 10 suppressed as equivalent, same
as the single-file version) exhaustively, and 74/79 (100%, 5 suppressed)
by default.
